added useSetCityLocation and useSetMarkers hooks
