+++ src/hooks/index.ts
@@ -1,5 +1,9 @@
 import useMap from './use-map.ts';
+import useSetCityLocation from './use-set-city-location.ts';
+import useSetMarkers from './use-set-markers.ts';
 
 export {
   useMap,
+  useSetCityLocation,
+  useSetMarkers,
 };
